Rename SemanticLabelClassifier to SemanticPixelClassifier
